+++ docs/changelog.rst
@@ -6,8 +6,9 @@ Changelog
 
 Upgrade steps from 1.9:
 
+* Install ``yarn`` (e.g. ``sudo npm install -g yarn``)
 * Update python libraries ``pip install -r requirements.txt``
-* Update CSS and JS libraries ``python manage.py bower install``
+* Update CSS and JS libraries ``yarn install``
 * Update static files (only production): ``python manage.py collectstatic``
 
 
@@ -25,6 +26,7 @@ Upgrade steps from 1.9:
 
 * Improved docker and docker-compose images `#340`_
 * Updated many libraries to last version (bootstrap, font awesome, etc.)
+* Use yarn to download CSS/JS libraries
 
 .. _#340: https://github.com/wger-project/wger/issues/340